Add NetworkModule.kt to android_core
