+++ android_core/src/main/kotlin/com/walletconect/android_core/di/CommonModule.kt
@@ -1,0 +1,1 @@
+package com.walletconect.android_core.di
