+++ android_core/src/main/kotlin/com/walletconect/android_core/network/RelayConnectionInterface.kt
@@ -1,0 +1,10 @@
+@file:JvmSynthetic
+
+package com.walletconect.android_core.network
+
+import com.walletconnect.foundation.network.RelayInterface
+
+interface RelayConnectionInterface : RelayInterface {
+    fun connect(onError: (String) -> Unit)
+    fun disconnect(onError: (String) -> Unit)
+}
