+++ android_core/src/main/kotlin/com/walletconect/android_core/network/data/connection/ConnectionEvent.kt
@@ -1,4 +1,4 @@
-package com.walletconnect.sign.network.data.connection
+package com.walletconect.android_core.network.data.connection
 
 internal enum class ConnectionEvent {
     CONNECT, DISCONNECT
