+++ android_core/src/main/kotlin/com/walletconect/android_core/network/data/connection/ConnectionType.kt
@@ -1,0 +1,5 @@
+package com.walletconect.android_core.network.data.connection
+
+internal enum class ConnectionType {
+    AUTOMATIC, MANUAL
+}
