+++ foundation/src/main/java/com/walletconnect/foundation/di/NetworkModule.kt
