+++ foundation/src/main/java/com/walletconnect/foundation/network/data/adapter/FlowStreamAdapter.kt
@@ -11,14 +11,14 @@ import kotlinx.coroutines.reactive.collect
 import java.lang.reflect.Type
 
 class FlowStreamAdapter<T> : StreamAdapter<T, Flow<T>> {
-    
+
     override fun adapt(stream: Stream<T>) = flow {
         stream.collect {
             emit(it)
         }
     }
 
-    internal class Factory : StreamAdapter.Factory {
+    class Factory : StreamAdapter.Factory {
 
         override fun create(type: Type): StreamAdapter<Any, Any> {
             return when (type.getRawType()) {
