+++ signSDK/sign/src/main/kotlin/com/walletconnect/sign/client/mapper/ClientMapper.kt
@@ -2,14 +2,11 @@ package com.walletconnect.sign.client.mapper
 
 import android.net.Uri
 import android.os.Build
-import android.util.Log
-import com.walletconnect.sign.BuildConfig
 import com.walletconnect.sign.client.Sign
 import com.walletconnect.sign.core.model.vo.jsonRpc.JsonRpcResponseVO
 import com.walletconnect.sign.core.model.vo.sync.PendingRequestVO
 import com.walletconnect.sign.engine.model.EngineDO
-import com.walletconnect.sign.network.data.connection.ConnectionType
-import java.net.URI
+import com.walletconect.android_core.network.data.connection.ConnectionType
 
 //TODO: Figure out what to do with models separation
 @JvmSynthetic
