--- signSDK/sign/src/main/kotlin/com/walletconnect/sign/network/data/connection/ConnectionType.kt
@@ -1,5 +1,0 @@
-package com.walletconnect.sign.network.data.connection
-
-internal enum class ConnectionType {
-    AUTOMATIC, MANUAL
-}
