+++ signSDK/sign/src/test/kotlin/com/walletconect/android_core/network/domain/relay/RelayRepositoryTest.kt
@@ -1,4 +1,4 @@
-package com.walletconnect.sign.network.domain.relay
+package com.walletconect.android_core.network.domain.relay
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
